chore: rename `fs.ts` to `resolver.ts`
