+++ packages/slidev/node/common.ts
@@ -4,7 +4,7 @@ import { loadConfigFromFile, mergeConfig, resolveConfig } from 'vite'
 import type { ConfigEnv, InlineConfig } from 'vite'
 import type { ResolvedSlidevOptions } from './options'
 import { generateGoogleFontsUrl } from './utils'
-import { clientRoot, toAtFS } from './fs'
+import { clientRoot, toAtFS } from './resolver'
 
 export async function getIndexHtml({ roots, data }: ResolvedSlidevOptions): Promise<string> {
   let main = await fs.readFile(join(clientRoot, 'index.html'), 'utf-8')
