+++ packages/slidev/node/plugins/setupNode.ts
@@ -3,7 +3,7 @@ import { fileURLToPath } from 'node:url'
 import fs from 'fs-extra'
 import { deepMergeWithArray } from '@antfu/utils'
 import jiti from 'jiti'
-import { clientRoot } from '../fs'
+import { clientRoot } from '../resolver'
 
 export async function loadSetups<T, R extends object>(
   roots: string[],
