+++ packages/slidev/node/plugins/unocss.ts
@@ -8,7 +8,7 @@ import { mergeConfigs } from 'unocss'
 import jiti from 'jiti'
 import UnoCSS from 'unocss/vite'
 import type { ResolvedSlidevOptions, SlidevPluginOptions } from '../options'
-import { clientRoot, userRoot } from '../fs'
+import { clientRoot, userRoot } from '../resolver'
 import { loadSetups } from './setupNode'
 
 export async function createUnocssPlugin(
